+++ apps/website/src/pages/_app.tsx
@@ -22,19 +22,16 @@ const MyApp = ({ Component, pageProps }: AppProps) => {
             footer={pageProps.footer !== false ? <Footer /> : null}
           >
             <Script
-              id="woopra-js"
+              id="plausible"
               strategy="afterInteractive"
+              data-domain="saas-ui.dev"
+              src="https://plausible.io/js/plausible.js"
+            />
+            <Script
+              id="plausible-snippet"
               dangerouslySetInnerHTML={{
-                __html: `!function(){var t,o,c,e=window,n=document,r=arguments,a="script",i=["call","cancelAction","config","identify","push","track","trackClick","trackForm","update","visit"],s=function(){var t,o=this,c=function(t){o[t]=function(){return o._e.push([t].concat(Array.prototype.slice.call(arguments,0))),o}};for(o._e=[],t=0;t<i.length;t++)c(i[t])};for(e.__woo=e.__woo||{},t=0;t<r.length;t++)e.__woo[r[t]]=e[r[t]]=e[r[t]]||new s;(o=n.createElement(a)).async=1,o.src="https://static.woopra.com/js/w.js",(c=n.getElementsByTagName(a)[0]).parentNode.insertBefore(o,c)}("woopra");
-
-  woopra.config({
-    domain: "saas-ui.dev",
-    outgoing_tracking: true,
-    download_tracking: true,
-    click_tracking: true
-  });
-
-  woopra.track();`,
+                __html:
+                  'window.plausible = window.plausible || function() { (window.plausible.q = window.plausible.q || []).push(arguments) }',
               }}
             />
             <Script