+++ packages/saas-ui-clerk/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @saas-ui/clerk
 
+## 1.1.3
+
+### Patch Changes
+
+- Updated dependencies [d003608]
+  - @saas-ui/auth@1.3.0
+
+## 1.1.2
+
+### Patch Changes
+
+- @saas-ui/auth@1.2.2
+
+## 1.1.1
+
+### Patch Changes
+
+- @saas-ui/auth@1.2.1
+
 ## 1.1.0
 
 ### Minor Changes