+++ packages/saas-ui-forms/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @saas-ui/forms
 
+## 1.2.0
+
+### Minor Changes
+
+- d003608: Form now accepts a render function with form state props.
+- d003608: Re-export all React Hook Form types and hooks
+- d003608: New onChange prop for Form that triggers when the form state is changed.
+
+## 1.1.2
+
+### Patch Changes
+
+- bef671d: Field variant now passed down to the internal Input.
+
+## 1.1.1
+
+### Patch Changes
+
+- @saas-ui/stepper@1.1.1
+
 ## 1.1.0
 
 ### Minor Changes