+++ packages/saas-ui-react/CHANGELOG.md
@@ -1,5 +1,49 @@
 # @saas-ui/react
 
+## 1.4.0
+
+### Minor Changes
+
+- d003608: Form now accepts a render function with form state props.
+- d003608: Re-export all React Hook Form types and hooks
+- d003608: New onChange prop for Form that triggers when the form state is changed.
+
+### Patch Changes
+
+- Updated dependencies [d003608]
+- Updated dependencies [d003608]
+- Updated dependencies [d003608]
+  - @saas-ui/auth@1.3.0
+  - @saas-ui/forms@1.2.0
+  - @saas-ui/modals@1.2.0
+
+## 1.3.3
+
+### Patch Changes
+
+- b7f169b: Removed fontSize from HotkeysListItem base styles.
+- 935a47c: useLocalStorage now returns the correct type.
+- bef671d: Field variant now passed down to the internal Input.
+- Updated dependencies [b7f169b]
+- Updated dependencies [935a47c]
+- Updated dependencies [bef671d]
+  - @saas-ui/hotkeys@1.1.1
+  - @saas-ui/hooks@1.1.1
+  - @saas-ui/forms@1.1.2
+  - @saas-ui/auth@1.2.2
+  - @saas-ui/modals@1.1.2
+
+## 1.3.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @saas-ui/collapse@1.1.0
+  - @saas-ui/stepper@1.1.1
+  - @saas-ui/forms@1.1.1
+  - @saas-ui/auth@1.2.1
+  - @saas-ui/modals@1.1.1
+
 ## 1.3.1
 
 ### Patch Changes