+++ packages/saas-ui-stepper/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @saas-ui/stepper
 
+## 1.1.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @saas-ui/collapse@1.1.0
+
 ## 1.1.0
 
 ### Minor Changes